requirements.txt
# Singapore Smart City Analytics — Dependencies
# Install: pip install -r requirements.txt

# Core ML & Computer Vision
ultralytics>=8.3.0          # YOLOv11 detection + training
torch>=2.1.0                # PyTorch backend
torchvision>=0.16.0         # Vision utilities
timm>=1.0.0                 # Pre-trained models (ResNet for context encoder)

# Tracking
boxmot>=11.0.0              # Multi-object tracking (BoT-SORT, StrongSORT)

# Data Collection & Processing
aiohttp>=3.9.0              # Async HTTP for parallel camera downloads
Pillow>=10.0.0              # Image handling

# Data Analysis
pandas>=2.1.0               # Tabular data handling
numpy>=1.24.0               # Numerical operations

# Configuration & CLI
pyyaml>=6.0                 # YAML config parsing
click>=8.1.0                # CLI interface

# Monitoring & Visualization
scipy>=1.11.0               # Statistical tests (KS test, PSI)
matplotlib>=3.8.0           # Plotting for reports

# API
fastapi>=0.108.0            # REST API server
uvicorn>=0.25.0             # ASGI server

# MLOps
mlflow>=2.9.0               # Experiment tracking

# Testing
pytest>=7.4.0               # Test framework
pytest-asyncio>=0.23.0      # Async test support
responses>=0.24.0           # Mock HTTP responses
